Start implementing MCR + once daily image check
