+++ src/tree/LocalRootTreeItemBase.ts
@@ -400,8 +400,8 @@ export abstract class LocalRootTreeItemBase<TItem extends DockerObject, TPropert
 }
 
 export function getTreeId(object: DockerObject): string {
-    // Id and State aren't defined for all Docker objects, but the concatenation of whatever exists of the three of these is enough to always be unique
+    // Several of these aren't defined for all Docker objects, but the concatenation of whatever exists among them is enough to always be unique
     // *and* change the ID when the state of the object changes
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
-    return `${object.Id}${object.Name}${(object as any).State}${(object as any).Current}`;
+    return `${object.Id}${object.Name}${(object as any).State}${(object as any).Current}${(object as any).Outdated}`;
 }
